khaoula
khaoula

+++ backend-laravel/database/migrations/2025_07_28_101745_remove_taille_from_materiels_table.php
@@ -9,10 +9,12 @@ return new class extends Migration
     /**
      * Run the migrations.
      */
-    public function up(): void
+  public function up(): void
 {
     Schema::table('materiels', function (Blueprint $table) {
-        $table->dropColumn('taille');
+        if (Schema::hasColumn('materiels', 'taille')) {
+            $table->dropColumn('taille');
+        }
     });
 }
